Created Trie.cpp
This program implements trie for binary numbers. Currently it possess only insert and search of numbers in O(16) as it is for 16 bit numbers. But it can be easily modified to 32 bit trie...Soon i will be posting generalized trie with many operations in one code and it will work even for strings too.